Fix/properly place register button (#734)
* Show small logo on mobile, always use svg logo

* Language Select: Remove "onhover" prop on mobile

* Show small logo on small screens

* Split HubsSubHeader in multiple components and make it mobile friendly

* remove unnecessary console.log

* update import path for HubsSubHeader

* Use Beta Logo again

* run 'yarn format'
