+++ frontend/pages/browse.js
@@ -16,7 +16,7 @@ import { nullifyUndefinedValues } from "../public/lib/profileOperations";
 import BrowseContent from "../src/components/browse/BrowseContent";
 import UserContext from "../src/components/context/UserContext";
 import TopOfPage from "../src/components/hooks/TopOfPage";
-import HubsSubHeader from "../src/components/indexPage/HubsSubHeader";
+import HubsSubHeader from "../src/components/indexPage/hubsSubHeader/HubsSubHeader";
 import MainHeadingContainerMobile from "../src/components/indexPage/MainHeadingContainerMobile";
 import WideLayout from "../src/components/layouts/WideLayout";
