+++ frontend/src/components/header/Header.js
@@ -310,6 +310,7 @@ export default function Header({
   const texts = getTexts({ page: "navigation", locale: locale });
   const [anchorEl, setAnchorEl] = React.useState(false);
   const isNarrowScreen = useMediaQuery((theme) => theme.breakpoints.down("xs"));
+  const isMediumScreen = useMediaQuery((theme) => theme.breakpoints.down("sm"));
   const LINKS = getLinks(pathName, texts);
   const toggleShowNotifications = (event) => {
     if (!anchorEl) setAnchorEl(event.currentTarget);
@@ -319,6 +320,15 @@ export default function Header({
 
   const onNotificationsClose = () => setAnchorEl(null);
 
+  const getLogo = () => {
+    if (isMediumScreen) {
+      return transparentHeader ? "/images/logo_white_no_text.svg" : "/images/logo_no_text.svg";
+    }
+    return transparentHeader ? "/images/logo_white_beta.svg" : "/images/logo.png";
+  };
+
+  const logo = getLogo();
+
   return (
     <Box
       component="header"
@@ -326,11 +336,7 @@ export default function Header({
     >
       <Container className={classes.container}>
         <Link href={localePrefix + "/"}>
-          <img
-            src={transparentHeader ? "/images/logo_white_beta.svg" : "/images/logo.png"}
-            alt={texts.climate_connect_logo}
-            className={classes.logo}
-          />
+          <img src={logo} alt={texts.climate_connect_logo} className={classes.logo} />
         </Link>
         {isNarrowScreen ? (
           <NarrowScreenLinks
@@ -520,7 +526,6 @@ function NarrowScreenLinks({
       !(loggedInUser && link.onlyShowLoggedOut) &&
       !(!loggedInUser && link.onlyShowLoggedIn)
   );
-  console.log(linksOutsideDrawer);
   return (
     <>
       <Box>
