+++ frontend/src/components/header/LanguageSelect.js
@@ -1,9 +1,8 @@
 import { Button, makeStyles, useMediaQuery } from "@material-ui/core";
 import LanguageIcon from "@material-ui/icons/Language";
 import { useRouter } from "next/router";
-import Cookies from "universal-cookie";
 import React, { useContext, useEffect, useRef, useState } from "react";
-
+import Cookies from "universal-cookie";
 import { getCookieProps } from "../../../public/lib/cookieOperations";
 import theme from "../../themes/theme";
 import UserContext from "../context/UserContext";
@@ -35,7 +34,8 @@ export default function LanguageSelect({ transparentHeader }) {
   const buttonRef = useRef(null);
   const [anchorEl, setAnchorEl] = useState(buttonRef.current);
   const [open, setOpen] = useState(false);
-  const isNarrowScreen = useMediaQuery(theme.breakpoints.down("sm"));
+  const isMediumScreen = useMediaQuery(theme.breakpoints.down("sm"));
+  const isNarrowScreen = useMediaQuery(theme.breakpoints.down("xs"));
   const router = useRouter();
   useEffect(function () {
     setAnchorEl(buttonRef.current);
@@ -72,13 +72,18 @@ export default function LanguageSelect({ transparentHeader }) {
     }
   };
 
+  const hoverButtonProps = {};
+
+  if (!isNarrowScreen) {
+    (hoverButtonProps.onMouseEnter = handleOpen), (hoverButtonProps.onMouseLeave = handleClose);
+  }
+
   // TODO: this could be generalized into a HoverButton component,
   // and used in the Welcome Blurb feature on /hubs
   return (
     <>
       <Button
-        onMouseEnter={handleOpen}
-        onMouseLeave={handleClose}
+        {...hoverButtonProps}
         className={classes.root}
         ref={buttonRef}
         aria-owns="language-select"
@@ -110,7 +115,7 @@ export default function LanguageSelect({ transparentHeader }) {
             key={index}
             className={classes.centerText}
             selected={l === locale}
-            dense={!isNarrowScreen}
+            dense={!isMediumScreen}
             onClick={(e) => handleLanguageClick(e, l)}
           >
             {l.toUpperCase()}
